# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TipEase API - Cashless tipping platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserProfileRole =
  (typeof UserProfileRole)[keyof typeof UserProfileRole];

export const UserProfileRole = {
  guest: "guest",
  host: "host",
  vendor: "vendor",
  admin: "admin",
} as const;

export type UserProfileLimitTier =
  (typeof UserProfileLimitTier)[keyof typeof UserProfileLimitTier];

export const UserProfileLimitTier = {
  standard: "standard",
  verified: "verified",
  enhanced: "enhanced",
  custom: "custom",
} as const;

export interface UserProfile {
  id: string;
  clerkId: string;
  email: string;
  name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  role?: UserProfileRole | null;
  limitTier: UserProfileLimitTier;
  walletBalance: number;
  /** @nullable */
  stripeCustomerId?: string | null;
  createdAt: string;
}

export interface UpdateUserProfileBody {
  name?: string;
  phone?: string;
}

export type SetUserRoleBodyRole =
  (typeof SetUserRoleBodyRole)[keyof typeof SetUserRoleBodyRole];

export const SetUserRoleBodyRole = {
  guest: "guest",
  host: "host",
  vendor: "vendor",
  admin: "admin",
} as const;

export interface SetUserRoleBody {
  role: SetUserRoleBodyRole;
}

export type WalletLimitTier =
  (typeof WalletLimitTier)[keyof typeof WalletLimitTier];

export const WalletLimitTier = {
  standard: "standard",
  verified: "verified",
  enhanced: "enhanced",
  custom: "custom",
} as const;

export interface Wallet {
  balance: number;
  dailyLimit: number;
  dailyUsed: number;
  limitTier: WalletLimitTier;
  autoReload: boolean;
  /** @nullable */
  autoReloadAmount?: number | null;
  /** @nullable */
  autoReloadThreshold?: number | null;
}

export interface AddFundsBody {
  /** Amount in dollars */
  amount: number;
}

export interface PaymentIntentResponse {
  clientSecret: string;
  publishableKey: string;
  amount: number;
}

export type WalletTransactionType =
  (typeof WalletTransactionType)[keyof typeof WalletTransactionType];

export const WalletTransactionType = {
  deposit: "deposit",
  tip: "tip",
  withdrawal: "withdrawal",
  refund: "refund",
} as const;

export interface WalletTransaction {
  id: number;
  type: WalletTransactionType;
  amount: number;
  description: string;
  createdAt: string;
}

export type DailyLimitStatusTier =
  (typeof DailyLimitStatusTier)[keyof typeof DailyLimitStatusTier];

export const DailyLimitStatusTier = {
  standard: "standard",
  verified: "verified",
  enhanced: "enhanced",
  custom: "custom",
} as const;

export interface DailyLimitStatus {
  tier: DailyLimitStatusTier;
  dailyLimit: number;
  used: number;
  remaining: number;
  resetsAt: string;
}

export type EventEventType =
  (typeof EventEventType)[keyof typeof EventEventType];

export const EventEventType = {
  private_party: "private_party",
  wedding: "wedding",
  corporate: "corporate",
  restaurant: "restaurant",
  public_event: "public_event",
  other: "other",
} as const;

export type EventStatus = (typeof EventStatus)[keyof typeof EventStatus];

export const EventStatus = {
  draft: "draft",
  pending_vendor: "pending_vendor",
  confirmed: "confirmed",
  active: "active",
  ended: "ended",
  cancelled: "cancelled",
} as const;

export interface Event {
  id: string;
  hostId: string;
  /** @nullable */
  vendorId?: string | null;
  name: string;
  eventType: EventEventType;
  venueName: string;
  venueAddress: string;
  startTime: string;
  endTime: string;
  /** @nullable */
  expectedGuests?: number | null;
  /** @nullable */
  perGuestCap?: number | null;
  status: EventStatus;
  vendorConfirmed: boolean;
  hostName: string;
  /** @nullable */
  vendorName?: string | null;
  totalTips: number;
  tipCount: number;
  createdAt: string;
}

export type CreateEventBodyEventType =
  (typeof CreateEventBodyEventType)[keyof typeof CreateEventBodyEventType];

export const CreateEventBodyEventType = {
  private_party: "private_party",
  wedding: "wedding",
  corporate: "corporate",
  restaurant: "restaurant",
  public_event: "public_event",
  other: "other",
} as const;

export interface CreateEventBody {
  name: string;
  eventType: CreateEventBodyEventType;
  venueName: string;
  venueAddress: string;
  startTime: string;
  endTime: string;
  expectedGuests?: number;
  perGuestCap?: number;
  vendorId?: string;
}

export type UpdateEventBodyStatus =
  (typeof UpdateEventBodyStatus)[keyof typeof UpdateEventBodyStatus];

export const UpdateEventBodyStatus = {
  draft: "draft",
  pending_vendor: "pending_vendor",
  confirmed: "confirmed",
  active: "active",
  ended: "ended",
  cancelled: "cancelled",
} as const;

export interface UpdateEventBody {
  name?: string;
  status?: UpdateEventBodyStatus;
  perGuestCap?: number;
}

export interface ConfirmEventBody {
  confirmed: boolean;
}

export type QrCodeStatus = (typeof QrCodeStatus)[keyof typeof QrCodeStatus];

export const QrCodeStatus = {
  active: "active",
  revoked: "revoked",
  expired: "expired",
} as const;

export interface QrCode {
  id: string;
  eventId: string;
  token: string;
  qrCodeUrl: string;
  status: QrCodeStatus;
  validFrom: string;
  validUntil: string;
  scanCount: number;
  createdAt: string;
}

export interface QrValidationResult {
  valid: boolean;
  eventId: string;
  eventName: string;
  hostName: string;
  vendorName: string;
  venueName: string;
  eventType: string;
  /** @nullable */
  perGuestCap?: number | null;
  validUntil: string;
  message: string;
}

export interface SendTipBody {
  eventId: string;
  qrToken: string;
  amount: number;
}

export type TipStatus = (typeof TipStatus)[keyof typeof TipStatus];

export const TipStatus = {
  pending: "pending",
  completed: "completed",
  failed: "failed",
  refunded: "refunded",
} as const;

export interface Tip {
  id: string;
  eventId: string;
  guestId: string;
  vendorId: string;
  amount: number;
  eventName: string;
  vendorName: string;
  status: TipStatus;
  createdAt: string;
}

export type VendorStatus = (typeof VendorStatus)[keyof typeof VendorStatus];

export const VendorStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Vendor {
  id: string;
  userId: string;
  companyName: string;
  contactName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  businessRegNumber?: string | null;
  status: VendorStatus;
  /** @nullable */
  rating?: number | null;
  totalEvents: number;
  totalEarnings: number;
  createdAt: string;
}

export interface CreateVendorBody {
  companyName: string;
  contactName: string;
  phone?: string;
  businessRegNumber?: string;
}

export interface Notification {
  id: number;
  userId: string;
  type: string;
  message: string;
  read: boolean;
  createdAt: string;
}

export interface GuestDashboard {
  walletBalance: number;
  dailyUsed: number;
  dailyLimit: number;
  dailyRemaining: number;
  totalTipsSent: number;
  tipCount: number;
  recentTips: Tip[];
  unreadNotifications: number;
}

export interface HostDashboard {
  totalEvents: number;
  activeEvents: number;
  totalTipsGenerated: number;
  recentEvents: Event[];
  pendingVendorConfirmations: number;
  unreadNotifications: number;
}

export interface VendorDashboard {
  totalEarnings: number;
  pendingPayouts: number;
  todayEarnings: number;
  totalEvents: number;
  pendingEvents: number;
  recentTips: Tip[];
  /** @nullable */
  rating?: number | null;
  unreadNotifications: number;
}

export interface StripeKeyResponse {
  publishableKey: string;
}

export type ListVendorsParams = {
  q?: string;
};