# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TipEase API - Cashless tipping platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events for the current user (as host or vendor)
      security:
        - clerkAuth: []
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event (Host only)
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventBody"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{eventId}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event details
      security:
        - clerkAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Event not found
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      security:
        - clerkAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventBody"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{eventId}/confirm:
    post:
      operationId: confirmEventVendor
      tags: [events]
      summary: Vendor confirms or declines participation in an event
      security:
        - clerkAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmEventBody"
      responses:
        "200":
          description: Confirmation recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{eventId}/qr-code:
    post:
      operationId: generateQrCode
      tags: [qr-codes]
      summary: Generate a QR code for an event
      security:
        - clerkAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: QR code created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrCode"
    get:
      operationId: getEventQrCode
      tags: [qr-codes]
      summary: Get the QR code for an event
      security:
        - clerkAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: QR code info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrCode"
  /events/{eventId}/revoke:
    post:
      operationId: revokeEventQrCode
      tags: [qr-codes]
      summary: Revoke/pause a QR code
      security:
        - clerkAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: QR code revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrCode"
...
  /qr/validate/{token}:
    get:
      operationId: validateQrCode
      tags: [qr-codes]
      summary: Validate a QR code token and get event info (public endpoint)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: QR code is valid, returns event info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrValidationResult"
        "400":
          description: Invalid or expired QR code

  # Tips
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: Get notifications for current user
      security:
        - clerkAuth: []
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{notificationId}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      security:
        - clerkAuth: []
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"

  # Dashboard summaries (wow endpoints)
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    Event:
      type: object
      properties:
        id:
          type: string
        hostId:
          type: string
        vendorId:
          type: ["string", "null"]
        name:
          type: string
        eventType:
          type: string
          enum: [private_party, wedding, corporate, restaurant, public_event, other]
        venueName:
          type: string
        venueAddress:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        expectedGuests:
          type: ["integer", "null"]
        perGuestCap:
          type: ["number", "null"]
        status:
          type: string
          enum: [draft, pending_vendor, confirmed, active, ended, cancelled]
        vendorConfirmed:
          type: boolean
        hostName:
          type: string
        vendorName:
          type: ["string", "null"]
        totalTips:
          type: number
        tipCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - hostId
        - name
        - eventType
        - venueName
        - venueAddress
        - startTime
        - endTime
        - status
        - vendorConfirmed
        - hostName
        - totalTips
        - tipCount
        - createdAt
    CreateEventBody:
      type: object
      properties:
        name:
          type: string
        eventType:
          type: string
          enum: [private_party, wedding, corporate, restaurant, public_event, other]
        venueName:
          type: string
        venueAddress:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        expectedGuests:
          type: integer
        perGuestCap:
          type: number
        vendorId:
          type: string
      required:
        - name
        - eventType
        - venueName
        - venueAddress
        - startTime
        - endTime
    UpdateEventBody:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [draft, pending_vendor, confirmed, active, ended, cancelled]
        perGuestCap:
          type: number
    ConfirmEventBody:
      type: object
      properties:
        confirmed:
          type: boolean
      required:
        - confirmed
    QrCode:
      type: object
      properties:
        id:
          type: string
        eventId:
          type: string
        token:
          type: string
        qrCodeUrl:
          type: string
        status:
          type: string
          enum: [active, revoked, expired]
        validFrom:
          type: string
          format: date-time
        validUntil:
          type: string
          format: date-time
        scanCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - eventId
        - token
        - qrCodeUrl
        - status
        - validFrom
        - validUntil
        - scanCount
        - createdAt
    QrValidationResult:
      type: object
      properties:
        valid:
          type: boolean
        eventId:
          type: string
        eventName:
          type: string
        hostName:
          type: string
        vendorName:
          type: string
        venueName:
          type: string
        eventType:
          type: string
        perGuestCap:
          type: ["number", "null"]
        validUntil:
          type: string
          format: date-time
        message:
          type: string
      required:
        - valid
        - eventId
        - eventName
        - hostName
        - vendorName
        - venueName
        - eventType
        - validUntil
        - message
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        type:
          type: string
        message:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - type
        - message
        - read
        - createdAt
...