# lib/api-zod/src/generated/types/qrValidationResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TipEase API - Cashless tipping platform
 * OpenAPI spec version: 0.1.0
 */

export interface QrValidationResult {
  valid: boolean;
  eventId: string;
  eventName: string;
  hostName: string;
  vendorName: string;
  venueName: string;
  eventType: string;
  /** @nullable */
  perGuestCap?: number | null;
  validUntil: Date;
  message: string;
}